requirements.txt
# Core runtime
keyring
PyYAML
requests
termcolor

#memory storage
chromaDB

# Audio / STT (RealtimeSTT dependency)
RealtimeSTT

# Text-to-Speech (Piper)
piper-tts
sounddevice
numpy


